requirements.txt
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.2.2
sentence-transformers==3.3.1
torch==2.5.1+cu121
tqdm==4.67.1
wandb==0.19.1
pyyaml